Refactoring: GeoDataRiverSurvey::MoveRiverPathPointCommand is moved to separated files.
