+++ libs/geodata/riversurvey/geodatariverpathpointmovedialog.cpp
@@ -4,6 +4,7 @@
 #include "geodatariverpathpointmovedialog.h"
 #include "geodatariversurvey.h"
 #include "geodatariversurveybackgroundgridcreatethread.h"
+#include "private/geodatariversurvey_moveriverpathpointcommand.h"
 
 #include <misc/iricundostack.h>
 
@@ -51,60 +52,6 @@ void GeoDataRiverPathPointMoveDialog::setSingleSelection(bool single)
 		ui->coordY->clear();
 	}
 }
-
-class GeoDataRiverSurvey::MoveRiverPathPointCommand : public QUndoCommand
-{
-public:
-	MoveRiverPathPointCommand(bool apply, QVector2D offset, GeoDataRiverSurvey* rs) :
-		QUndoCommand {GeoDataRiverSurvey::tr("Move Traversal Lines")}
-	{
-		m_apply = apply;
-		GeoDataRiverPathPoint* p = rs->headPoint();
-		p = p->nextPoint();
-		while (p != 0) {
-			if (p->IsSelected) {
-				m_points.append(p);
-				m_oldPositions.append(p->position());
-				m_newPositions.append(p->position() + offset);
-			}
-			p = p->nextPoint();
-		}
-		m_rs = rs;
-	}
-	void undo() {
-		m_rs->m_gridThread->cancel();
-
-		for (int i = 0; i < m_points.count(); ++i) {
-			QVector2D oldpos = m_oldPositions.at(i);
-			m_points[i]->setPosition(oldpos);
-		}
-		if (! m_apply) {
-			m_rs->headPoint()->updateRiverShapeInterpolators();
-			m_rs->updateShapeData();
-			m_rs->renderGraphicsView();
-			m_rs->updateCrossectionWindows();
-		}
-	}
-	void redo() {
-		m_rs->m_gridThread->cancel();
-
-		for (int i = 0; i < m_points.count(); ++i) {
-			QVector2D newpos = m_newPositions.at(i);
-			m_points[i]->setPosition(newpos);
-		}
-		m_rs->headPoint()->updateRiverShapeInterpolators();
-		m_rs->setModified();
-		m_rs->updateShapeData();
-		m_rs->renderGraphicsView();
-		m_rs->updateCrossectionWindows();
-	}
-private:
-	bool m_apply;
-	QList<GeoDataRiverPathPoint*> m_points;
-	QList<QVector2D> m_oldPositions;
-	QList<QVector2D> m_newPositions;
-	GeoDataRiverSurvey* m_rs;
-};
 
 void GeoDataRiverPathPointMoveDialog::accept()
 {
