+++ libs/geodata/riversurvey/riversurvey.pro
@@ -124,6 +124,7 @@ HEADERS += gd_riversurvey_global.h \
            private/geodatariversurvey_mouseeditcrosssectioncommand.h \
            private/geodatariversurvey_mouserotaterivercrosssectioncommand.h \
            private/geodatariversurvey_mouseshiftriverpathcentercommand.h \
+           private/geodatariversurvey_moveriverpathpointcommand.h \
            private/geodatariversurvey_setdisplaysettingcommand.h \
            private/geodatariversurveycrosssectionwindow_impl.h \
            private/geodatariversurveyproxy_setsettingcommand.h
@@ -174,6 +175,7 @@ SOURCES += geodatarivercrosssection.cpp \
            private/geodatariversurvey_mouseeditcrosssectioncommand.cpp \
            private/geodatariversurvey_mouserotaterivercrosssectioncommand.cpp \
            private/geodatariversurvey_mouseshiftriverpathcentercommand.cpp \
+           private/geodatariversurvey_moveriverpathpointcommand.cpp \
            private/geodatariversurvey_setdisplaysettingcommand.cpp \
            private/geodatariversurveyproxy_setsettingcommand.cpp
 RESOURCES += riversurvey.qrc
